+++ src/main/java/com/stratio/qa/specs/GosecSpec.java
@@ -965,7 +965,7 @@ public class GosecSpec extends BaseGSpec {
                     // Add gids array to new json for PUT request
                     putObject.put("gids", stringGroups);
 
-                    commonspec.getLogger().warn("Json for PUT request---> {}", putObject.toString());
+                    commonspec.getLogger().warn("Json for PUT request---> {}", putObject);
                     Future<Response> response = commonspec.generateRequest("PUT", false, null, null, endPointGetProfile, JsonValue.readHjson(putObject.toString()).toString(), "json", "");
                     commonspec.setResponse("PUT", response.get());
                     if (commonspec.getResponse().getStatusCode() != 204) {
@@ -1111,7 +1111,7 @@ public class GosecSpec extends BaseGSpec {
         String resourcePrefix = getResourcePrefix(resource);
         String endPointPolicies = "/baas/management/policies";
         String endPoint = getResourceEndpoint(resource);
-        String endPointResource = endPoint + resourcePrefix + resourceId;;
+        String endPointResource = endPoint + resourcePrefix + resourceId;
 
 
         if (tenantOrig != null) {
@@ -1219,7 +1219,7 @@ public class GosecSpec extends BaseGSpec {
                             }
                         }
                         putObject.put(uidOrGidTenant, new JSONArray(stringGroups));
-                        commonspec.getLogger().debug("Json for PATCH request---> {}", putObject.toString());
+                        commonspec.getLogger().debug("Json for PATCH request---> {}", putObject);
                         Future<Response> response = commonspec.generateRequest("PATCH", false, null, null, endPointTenant, JsonValue.readHjson(putObject.toString()).toString(), "json", "");
                         commonspec.setResponse("PATCH", response.get());
                         if (commonspec.getResponse().getStatusCode() != 204) {
@@ -1250,21 +1250,27 @@ public class GosecSpec extends BaseGSpec {
      * @param keytab          : creation of keytab in vault (OPTIONAL)
      * @param certificate     : creation of certificate in vault (OPTIONAL)
      * @param groups          : groups for custom user (OPTIONAL)
+     * @param doesNotExist    : (IF false) resource will be overwritten if exists previously  (OPTIONAL)
      * @throws Exception
      */
-    @When("^I create custom user '(.+?)'( in tenant '(.+?)')?( with tenant user and tenant password '(.+:.+?)')?( generating keytab)?( generating certificate)?( assigned to groups '(.+?)')?$")
-    public void createUserResource(String userName, String tenantOrig, String tenantLoginInfo, String keytab, String certificate, String groups) throws Exception {
+    @When("^I create custom user '(.+?)'( in tenant '(.+?)')?( with tenant user and tenant password '(.+:.+?)')?( generating keytab)?( generating certificate)?( assigned to groups '(.+?)')?( if it does not exist)?$")
+    public void createUserResource(String userName, String tenantOrig, String tenantLoginInfo, String keytab, String certificate, String groups, String doesNotExist) throws Exception {
+        Boolean booleanExist = false;
+        if (doesNotExist != null) {
+            booleanExist = true;
+        }
         if (ThreadProperty.get("isKeosEnv") != null && ThreadProperty.get("isKeosEnv").equals("true")) {
-            createUserResourceKeos(userName, tenantOrig, tenantLoginInfo, keytab, certificate, groups);
+            createUserResourceKeos(userName, tenantOrig, tenantLoginInfo, keytab, certificate, groups, booleanExist);
         } else {
-            createUserResourceDcos(userName, tenantOrig, tenantLoginInfo, keytab, certificate, groups);
+            createUserResourceDcos(userName, tenantOrig, tenantLoginInfo, keytab, certificate, groups, booleanExist);
         }
     }
 
-    private void createUserResourceDcos(String userName, String tenantOrig, String tenantLoginInfo, String keytab, String certificate, String groups) throws Exception {
+    private void createUserResourceDcos(String userName, String tenantOrig, String tenantLoginInfo, String keytab, String certificate, String groups, boolean doesNotExist) throws Exception {
         String managementVersion = ThreadProperty.get("gosec-management_version");
         String managementBaasVersion = ThreadProperty.get("gosec-management-baas_version");
         String endPoint = "/service/gosecmanagement" + ThreadProperty.get("API_USER");
+        Integer[] expectedStatusDelete = {200, 204};
         Boolean addSourceType = false;
         Boolean managementBaas = false;
         String endPointResource = "";
@@ -1322,9 +1328,22 @@ public class GosecSpec extends BaseGSpec {
                 // Send POST request
                 sendIdentitiesPostRequest("user", userName, data, newEndPoint);
 
-
             } else {
-                commonspec.getLogger().warn("Custom user:{} already exist", userName);
+                if (doesNotExist) {
+                    commonspec.getLogger().warn("Custom user:{} already exist", userName);
+                } else {
+                    // Delete user and send the request again
+                    deleteResourceIfExistsDcos("user", userName, tenantOrig, tenantLoginInfo, endPoint, null);
+
+                    try {
+                        assertThat(commonspec.getResponse().getStatusCode()).isIn(expectedStatusDelete);
+                    } catch (AssertionError e) {
+                        commonspec.getLogger().warn("Error deleting User {}: {}", userName, commonspec.getResponse().getResponse());
+                        throw e;
+                    }
+                    //send request again
+                    createUserResourceDcos(userName, tenantOrig, tenantLoginInfo, keytab, certificate, groups, doesNotExist);
+                }
             }
 
         } catch (Exception e) {
@@ -1334,9 +1353,10 @@ public class GosecSpec extends BaseGSpec {
 
     }
 
-    private void createUserResourceKeos(String userName, String tenantOrig, String tenantLoginInfo, String keytab, String certificate, String groups) throws Exception {
+    private void createUserResourceKeos(String userName, String tenantOrig, String tenantLoginInfo, String keytab, String certificate, String groups, boolean doesNotExist) throws Exception {
         String endPoint = "/baas/management/user";
         String endPointResource = "";
+        Integer[] expectedStatusDelete = {200, 204};
         String uid = userName.replaceAll("\\s+", ""); //delete white spaces
         String data = generateBaasUserJson(uid, userName, groups, keytab, certificate);
         if (tenantOrig != null) {
@@ -1352,7 +1372,21 @@ public class GosecSpec extends BaseGSpec {
                 // Send POST request
                 sendIdentitiesPostRequest("user", userName, data, endPoint);
             } else {
-                commonspec.getLogger().warn("Custom user:{} already exist", userName);
+                if (doesNotExist) {
+                    commonspec.getLogger().warn("Custom user:{} already exist", userName);
+                } else {
+                    // Delete user and send the request again
+                    deleteResourceIfExistsKeos("user", userName, tenantOrig, tenantLoginInfo, endPoint, null);
+
+                    try {
+                        assertThat(commonspec.getResponse().getStatusCode()).isIn(expectedStatusDelete);
+                    } catch (AssertionError e) {
+                        commonspec.getLogger().warn("Error deleting User {}: {}", userName, commonspec.getResponse().getResponse());
+                        throw e;
+                    }
+                    //send request again
+                    createUserResourceKeos(userName, tenantOrig, tenantLoginInfo, keytab, certificate, groups, doesNotExist);
+                }
             }
         } catch (Exception e) {
             commonspec.getLogger().error("Rest Host or Rest Port are not initialized {}{}", commonspec.getRestHost(), commonspec.getRestPort());
@@ -1369,23 +1403,29 @@ public class GosecSpec extends BaseGSpec {
      * @param tenantLoginInfo : user and password to log into tenant (OPTIONAL)
      * @param users           : users for custom group (OPTIONAL)
      * @param groups          : groups for custom group (nested) (OPTIONAL)
+     * @param doesNotExist    : (IF false) resource will be overwritten if exists previously  (OPTIONAL)
      * @throws Exception
      */
-    @When("^I create custom group '(.+?)'( in tenant '(.+?)')?( with tenant user and tenant password '(.+:.+?)')?( assigned to users '(.+?)')?( assigned to groups '(.+?)')?$")
-    public void createGroupResource(String groupName, String tenantOrig, String tenantLoginInfo, String users, String groups) throws Exception {
+    @When("^I create custom group '(.+?)'( in tenant '(.+?)')?( with tenant user and tenant password '(.+:.+?)')?( assigned to users '(.+?)')?( assigned to groups '(.+?)')?( if it does not exist)?$")
+    public void createGroupResource(String groupName, String tenantOrig, String tenantLoginInfo, String users, String groups, String doesNotExist) throws Exception {
+        Boolean booleanExist = false;
+        if (doesNotExist != null) {
+            booleanExist = true;
+        }
         if (ThreadProperty.get("isKeosEnv") != null && ThreadProperty.get("isKeosEnv").equals("true")) {
-            createGroupResourceKeos(groupName, tenantOrig, tenantLoginInfo, users, groups);
+            createGroupResourceKeos(groupName, tenantOrig, tenantLoginInfo, users, groups, booleanExist);
         } else {
-            createGroupResourceDcos(groupName, tenantOrig, tenantLoginInfo, users, groups);
+            createGroupResourceDcos(groupName, tenantOrig, tenantLoginInfo, users, groups, booleanExist);
         }
     }
 
-    private void createGroupResourceDcos(String groupName, String tenantOrig, String tenantLoginInfo, String users, String groups) throws Exception {
+    private void createGroupResourceDcos(String groupName, String tenantOrig, String tenantLoginInfo, String users, String groups, boolean doesNotExist) throws Exception {
         String managementVersion = ThreadProperty.get("gosec-management_version");
         String managementBaasVersion = ThreadProperty.get("gosec-management-baas_version");
         String endPoint = "/service/gosecmanagement" + ThreadProperty.get("API_GROUP");
         Boolean addSourceType = false;
         Boolean managementBaas = false;
+        Integer[] expectedStatusDelete = {200, 204};
         String endPointResource = "";
         String gid = groupName.replaceAll("\\s+", ""); //delete white spaces
         Integer expectedStatusCreate = 201;
@@ -1443,7 +1483,21 @@ public class GosecSpec extends BaseGSpec {
                 sendIdentitiesPostRequest("group", groupName, data, newEndPoint);
 
             } else {
-                commonspec.getLogger().warn("Custom group:{} already exist", groupName);
+                if (doesNotExist) {
+                    commonspec.getLogger().warn("Custom group:{} already exist", groupName);
+                } else {
+                    // Delete group and send the request again
+                    deleteResourceIfExistsDcos("group", groupName, tenantOrig, tenantLoginInfo, endPoint, null);
+
+                    try {
+                        assertThat(commonspec.getResponse().getStatusCode()).isIn(expectedStatusDelete);
+                    } catch (AssertionError e) {
+                        commonspec.getLogger().warn("Error deleting Group {}: {}", groupName, commonspec.getResponse().getResponse());
+                        throw e;
+                    }
+                    //send request again
+                    createGroupResourceDcos(groupName, tenantOrig, tenantLoginInfo, users, groups, doesNotExist);
+                }
             }
 
         } catch (Exception e) {
@@ -1453,11 +1507,12 @@ public class GosecSpec extends BaseGSpec {
 
     }
 
-    private void createGroupResourceKeos(String groupName, String tenantOrig, String tenantLoginInfo, String users, String groups) throws Exception {
+    private void createGroupResourceKeos(String groupName, String tenantOrig, String tenantLoginInfo, String users, String groups, boolean doesNotExist) throws Exception {
         String endPoint = "/baas/management/group";
         String endPointResource = "";
         String gid = groupName.replaceAll("\\s+", ""); //delete white spaces
         String data = generateBaasGroupJson(gid, groupName, users, groups);
+        Integer[] expectedStatusDelete = {200, 204};
 
         if (tenantOrig != null) {
             // Set REST connection
@@ -1473,7 +1528,21 @@ public class GosecSpec extends BaseGSpec {
                 // Send POST request
                 sendIdentitiesPostRequest("group", groupName, data, endPoint);
             } else {
-                commonspec.getLogger().warn("Custom group:{} already exist", groupName);
+                if (doesNotExist) {
+                    commonspec.getLogger().warn("Custom group:{} already exist", groupName);
+                } else {
+                    // Delete group and send the request again
+                    deleteResourceIfExistsKeos("group", groupName, tenantOrig, tenantLoginInfo, endPoint, null);
+
+                    try {
+                        assertThat(commonspec.getResponse().getStatusCode()).isIn(expectedStatusDelete);
+                    } catch (AssertionError e) {
+                        commonspec.getLogger().warn("Error deleting Group {}: {}", groupName, commonspec.getResponse().getResponse());
+                        throw e;
+                    }
+                    //send request again
+                    createGroupResourceKeos(groupName, tenantOrig, tenantLoginInfo, users, groups, doesNotExist);
+                }
             }
         } catch (Exception e) {
             commonspec.getLogger().error("Rest Host or Rest Port are not initialized {}{}", commonspec.getRestHost(), commonspec.getRestPort());